Scripts for synchronizing and publishing boxes introduced
1) `sync_images` - Script for downloading images introduced

`sync_images` is the POSIX shell script which parses JSON configuration
files for various distributions and checks if ISO images are downloaded
and their checksums are correct.

It uses `jq` and `wget` utilities in order to parse JSON files and
download ISOs.

2) `publish_vm` - Script for publishing images on Vagrant Cloud added

`publish_vm` is the POSIX shell script which uses `vagrant` utility
to upload recently built boxes to Vagrant Cloud.